chore: refactor to modular structure
